requirements.txt
gradio
pandas
numpy
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
chromadb


